### 2021-11-12 #3913 @whedon
[PRE REVIEW]: CR-Sparse: Hardware accelerated functional algorithms for sparse signal processing in Python using JAX
**Submitting author:** @shailesh1729 (<a href="http://orcid.org/0000-0003-2217-4768">Shailesh Kumar</a>)
**Repository:** <a href="https://github.com/carnotresearch/cr-sparse" target ="_blank">https://github.com/carnotresearch/cr-sparse</a>
**Version:** v0.2.1
