### 2021-11-13 923012a
arr_l1norm
