+++ src/cr/sparse/__init__.py
@@ -218,6 +218,7 @@ from cr.sparse._src.types import (
 
 from cr.sparse._src.ndarray import (
     arr_largest_index,
+    arr_l1norm,
     arr_l2norm,
     arr_l2norm_sqr,
     arr_vdot,
