+++ paper/paper.bib
@@ -2,6 +2,8 @@
   title={A computational model for TensorFlow: an introduction},
   author={Abadi, Mart{\'\i}n and Isard, Michael and Murray, Derek G},
   booktitle={Proceedings of the 1st ACM SIGPLAN International Workshop on Machine Learning and Programming Languages},
+  url={https://doi.org/10.1145/3088525.3088527},
+  doi={10.1145/3088525.3088527},
   pages={1--7},
   year={2017}
 }
@@ -14,6 +16,8 @@
   number={2},
   pages={218--233},
   year={2003},
+  doi={10.1109/ICCV.2001.937651},
+  url = {https://doi.org/10.1109/iccv.2001.937651},
   publisher={IEEE}
 }
 
@@ -25,6 +29,8 @@
   number={3},
   pages={265--274},
   year={2009},
+  doi = {10.1016/j.acha.2009.04.002},
+	url = {https://doi.org/10.1016/j.acha.2009.04.002},
   publisher={Elsevier}
 }
 
@@ -36,6 +42,8 @@
   number={2},
   pages={298--309},
   year={2010},
+  doi={10.1109/jstsp.2010.2042411},
+	url={https://doi.org/10.1109/jstsp.2010.2042411},
   publisher={IEEE}
 }
 
@@ -46,6 +54,8 @@
   booktitle={Visual Motion, 1991., Proceedings of the IEEE Workshop on},
   pages={179--186},
   year={1991},
+  doi={10.1109/wvm.1991.212809},
+	url={https://doi.org/10.1109/wvm.1991.212809},
   organization={IEEE}
 }
 
@@ -57,6 +67,8 @@
   number={3},
   pages={159--179},
   year={1998},
+  doi={10.1023/A:1008000628999},
+  uril={https://doi.org/10.1023/A:1008000628999},
   publisher={Springer}
 }
 
@@ -68,6 +80,8 @@
   number={5},
   pages={2230--2249},
   year={2009},
+  doi={10.1109/tit.2009.2016006},
+	url={https://doi.org/10.1109/tit.2009.2016006},
   publisher={IEEE}
 }
 
@@ -79,6 +93,8 @@
   number={9},
   pages={4395--4401},
   year={2010},
+  doi = {10.1109/tit.2010.2054653},
+	url = {https://doi.org/10.1109/tit.2010.2054653},
   publisher={IEEE}
 }
 
@@ -91,7 +107,8 @@
   number={1},
   pages={2487--2517},
   year={2013},
-  publisher={JMLR. org}
+  url={http://jmlr.org/papers/v14/dyer13a.html},
+  publisher={JMLR.org}
 }
 
 
@@ -100,6 +117,8 @@
   title={Sparse and redundant representations},
   author={Elad, Michael},
   year={2010},
+  doi={10.1007/978-1-4419-7011-4},
+	url={https://doi.org/10.1007/978-1-4419-7011-4},
   publisher={Springer}
 }
 
@@ -109,6 +128,8 @@
   booktitle={Computer Vision and Pattern Recognition, 2009. CVPR 2009. IEEE Conference on},
   pages={2790--2797},
   year={2009},
+  doi={10.1109/cvprw.2009.5206547},
+	url={https://doi.org/10.1109/cvprw.2009.5206547},
   organization={IEEE}
 }
 
@@ -120,6 +141,8 @@
   number={11},
   pages={2765--2781},
   year={2013},
+  doi={10.1109/tpami.2013.57},
+	url={https://doi.org/10.1109/tpami.2013.57},
   publisher={IEEE}
 }
 
@@ -149,6 +172,8 @@
   volume={1},
   pages={I--11},
   year={2003},
+  doi={10.1109/CVPR.2003.1211332},
[…]
