### 2021-11-15 #3913 @pdebuyl
@saran-nns @mirca would you review the article "CR-Sparse: Hardware accelerated functional algorithms for sparse signal processing in Python using JAX" for the Journal of Open Source Software?
