### 2021-11-15 3b3a5a8
making sure that the arguments are arrays
