### 2021-11-15 4d19b4b
combined proximal value and vector operator
