+++ src/cr/sparse/_src/opt/proximal_ops/basic.py
@@ -1,0 +1,34 @@
+# Copyright 2021 CR.Sparse Development Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from jax import jit, lax
+
+import jax.numpy as jnp
+import cr.sparse as crs
+
+
+def prox_zero():
+
+    @jit
+    def func(x):
+        return 0.
+
+    @jit
+    def proximal_op(x, t):
+        x = jnp.asarray(x)
+        x = crs.promote_arg_dtypes(x)
+        return x
+
+    return func, proximal_op
