### 2021-11-15 f2e52bc
smooth functions module, smooth_constant
