+++ src/cr/sparse/opt/smooth.py
@@ -12,24 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-Optimization Utilities
-"""
-# pylint: disable=W0611
 
-
-from cr.sparse._src.opt.projections import (
-    project_to_ball,
-    project_to_box,
-    project_to_real_upper_limit
-)
-
-from cr.sparse._src.opt.shrinkage import (
-    shrink
+from cr.sparse._src.opt.smooth import (
+    smooth_value_grad
 )
 
 
-from .indicators import *
-from .projectors import *
-from .proximal_ops import *
-from .smooth import *
+from cr.sparse._src.opt.smooth.constant import (
+    smooth_constant
+)
