+++ tests/lop/test_calculus.py
@@ -48,17 +48,17 @@ def test_first_derivative_centered():
     AT = lop.to_adjoint_matrix(T)
     assert jnp.allclose(A.T, AT)
     e1 = jnp.array([  0. ,  1.5, -4. , -5.5,  6.5, -1.5,  3.5,  0.])
-    assert_allclose(T.times(x), e1)
-    assert_allclose(A @ x, e1)
+    assert_allclose(T.times(x), e1, atol=atol, rtol=rtol)
+    assert_allclose(A @ x, e1, atol=atol, rtol=rtol)
     e2 = jnp.array([ -1.5, -2. ,  4. ,  5.5, -6.5,  1.5,  4. , -5.])
-    assert_allclose(T.trans(x), e2)
-    assert_allclose(AT @ x, e2)
+    assert_allclose(T.trans(x), e2, atol=atol, rtol=rtol)
+    assert_allclose(AT @ x, e2, atol=atol, rtol=rtol)
     assert lop.dot_test_real(keys[0], T)
     d = jnp.array([-1, 1]) / 2.
     for i in range(1, n-1):
-        assert_allclose(A[i, i-1:i+2:2],d)
+        assert_allclose(A[i, i-1:i+2:2],d, atol=atol, rtol=rtol)
     for i in range(1, n-1):
-        assert_allclose(AT[i-1:i+2:2, i],d)
+        assert_allclose(AT[i-1:i+2:2, i],d, atol=atol, rtol=rtol)
 
 
 def test_second_derivative():
@@ -67,15 +67,15 @@ def test_second_derivative():
     T = lop.jit(lop.second_derivative(n))
     A = lop.to_matrix(T)
     AT = lop.to_adjoint_matrix(T)
-    assert jnp.allclose(A.T, AT)
+    assert_allclose(A.T, AT, atol=atol, rtol=rtol)
     d = jnp.array([1, -2, 1])
     for i in range(1, n-1):
-        assert_allclose(A[i, i-1:i+2],d)
+        assert_allclose(A[i, i-1:i+2],d, atol=atol, rtol=rtol)
     for i in range(1, n-1):
-        assert_allclose(AT[i-1:i+2, i],d)
+        assert_allclose(AT[i-1:i+2, i],d, atol=atol, rtol=rtol)
     e1 = jnp.array([0.,  -1., -10.,   7.,  17., -33.,  43.,   0])
     e2 = jnp.array([3.,  -2., -10.,   7.,  17., -33.,  28., -10.])
-    assert_allclose(T.times(x), e1)
-    assert_allclose(A @ x, e1)
-    assert_allclose(T.trans(x), e2)
-    assert_allclose(AT @ x, e2)
+    assert_allclose(T.times(x), e1, atol=atol, rtol=rtol)
+    assert_allclose(A @ x, e1, atol=atol, rtol=rtol)
+    assert_allclose(T.trans(x), e2, atol=atol, rtol=rtol)
+    assert_allclose(AT @ x, e2, atol=atol, rtol=rtol)
