+++ src/cr/sparse/__init__.py
@@ -224,6 +224,7 @@ from cr.sparse._src.ndarray import (
     arr_vdot,
     arr_rdot,
     arr2vec,
+    log_pos,
 )
 
 from cr.sparse._src.similarity import (
