+++ src/cr/sparse/opt/smooth.py
@@ -18,6 +18,7 @@ from cr.sparse._src.opt.smooth import (
 )
 
 
-from cr.sparse._src.opt.smooth.constant import (
-    smooth_constant
-)
+from cr.sparse._src.opt.smooth.constant import smooth_constant
+from cr.sparse._src.opt.smooth.entropy import (
+    smooth_entropy,
+    smooth_entropy_vg)
