### 2021-11-16 cfa80d7
smooth_linear
