### 2021-11-16 017f063
smooth_logdet
