### 2021-11-16 dbbc733
smooth_quad_matrix
