+++ src/cr/sparse/opt/smooth.py
@@ -25,3 +25,6 @@ from cr.sparse._src.opt.smooth.entropy import (
 from cr.sparse._src.opt.smooth.huber import smooth_huber
 from cr.sparse._src.opt.smooth.linear import smooth_linear
 from cr.sparse._src.opt.smooth.logdet import smooth_logdet
+from cr.sparse._src.opt.smooth.quad import (
+    smooth_quad_matrix
+)
