+++ src/cr/sparse/_src/signal.py
@@ -40,6 +40,13 @@ def find_first_signal_with_energy_le_cw(X, energy):
 
 def randomize_rows(key, X):
     """Randomizes the rows in X
+
+    Args:
+        key: a PRNG key used as the random key.
+        X (jax.numpy.ndarray): A 2D data matrix
+
+    Returns:
+        (jax.numpy.ndarray): The data matrix with randomized rows
     """
     assert is_matrix(X)
     m, n = X.shape
@@ -48,6 +55,13 @@ def randomize_rows(key, X):
 
 def randomize_cols(key, X):
     """Randomizes the columns in X
+
+    Args:
+        key: a PRNG key used as the random key.
+        X (jax.numpy.ndarray): A 2D data matrix
+
+    Returns:
+        (jax.numpy.ndarray): The data matrix with randomized columns
     """
     assert is_matrix(X)
     m, n = X.shape
@@ -57,34 +71,77 @@ def randomize_cols(key, X):
 
 def largest_indices(x, K):
     """Returns the indices of K largest entries in x by magnitude
+
+    Args:
+        x (jax.numpy.ndarray): An data vector/point
+        K (int): The number of largest entries to be identified in x
+
+    Returns:
+        (jax.numpy.ndarray): An index vector of size K identifying the K largest entries in x
+        in descending order
     """
     indices = jnp.argsort(jnp.abs(x))
     return indices[:-K-1:-1]
 
 def largest_indices_rw(X, K):
     """Returns the indices of K largest entries by magnitude in each row of X
+
+    Args:
+        X (jax.numpy.ndarray): An (S,N) data matrix with data points in rows
+        K (int): The number of largest entries to be identified in each row of X
+
+    Returns:
+        (jax.numpy.ndarray): An (S,K) index matrix indices of K largest elements in each row of X
     """
     indices = jnp.argsort(jnp.abs(X), axis=1)
     return indices[:, :-K-1:-1]
 
 def largest_indices_cw(X, K):
     """Returns the indices of K largest entries by magnitude in each column of X
+
+    Args:
+        X (jax.numpy.ndarray): An (N,S) data matrix with data points in columns
+        K (int): The number of largest entries to be identified in each column of X
+
+    Returns:
+        (jax.numpy.ndarray): An (K,S) index matrix indices of K largest elements in each column of X
     """
     indices = jnp.argsort(jnp.abs(X), axis=0)
     return indices[:-K-1:-1, :]
 
 def take_along_rows(X, indices):
     """Picks K entries from each row of X specified by indices matrix
+
+    Args:
+        X (jax.numpy.ndarray): An (S,N) data matrix with data points in rows
+        indices (jax.numpy.ndarray): An (S,K) index matrix identifying the values to be picked up from X
+
+    Returns:
+        (jax.numpy.ndarray): An (S,K) data matrix subset of X containing K elements from each row of X
     """
     return jnp.take_along_axis(X, indices, axis=1)
 
 def take_along_cols(X, indices):
     """Picks K entries from each column of X specified by indices matrix
+
+    Args:
+        X (jax.numpy.ndarray): An (N,S) data matrix with data points in columns
+        indices (jax.numpy.ndarray): An (K,S) index matrix identifying the values to be picked up from X
+
+    Returns:
+        (jax.numpy.ndarray): An (K,S) data matrix subset of X containing K elements from each column of X
     """
     return jnp.take_along_axis(X, indices, axis=0)
 
 def sparse_approximation(x, K):
     """Keeps only largest K non-zero entries by magnitude in a vector x
+
+    Args:
+        x (jax.numpy.ndarray): An data vector/point
+        K (int): The number of largest entries to be kept in x
+
[…]
