### 2021-11-16 dd20083
driver for smooth conic dual solver
