+++ src/cr/sparse/_src/opt/proximal_ops/__init__.py
@@ -12,15 +12,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from functools import partial
 from jax import jit
 
 def prox_value_vec(func, prox_op):
     """Returns a function which computes both proximal value and vector
     """
-    @jit
-    def operator(x, t):
-        x = prox_op(x, t)
-        v = func(x)
-        return v, x
+    @partial(jit, static_argnums=(2,))
+    def operator(x, t, mode=0):
+        """
+        mode=0 only function value at x
+        mode=1 only proximal vector for x
+        mode=2 both proximal vector for x and function value at the proximal vector
+        """
+        if mode == 0:
+            return func(x)
+        if mode == 1:
+            return prox_op(x, t)
+        if mode == 2:
+            x = prox_op(x, t)
+            v = func(x)
+            return v, x
+        return None
 
     return operator
