+++ src/cr/sparse/_src/focs/l1rls.py
@@ -22,7 +22,9 @@ from .defs import FOCSOptions
 def l1rls(A, b, lambda_, x0, options: FOCSOptions = FOCSOptions()):
     """Solver for l1 regulated least square problem
     """
-    f = opt.smooth_quad_matrix
-    h = opt.prox_l1(lambda_)
+    func, grad = opt.smooth_quad_matrix()
+    f = opt.smooth_value_grad(func, grad)
+    h_func, h_prox = opt.prox_l1(lambda_)
+    h = opt.prox_value_vec(h_func, h_prox)
     af = matrix_affine_func(A, -b)
     return focs(f, af, h, x0, options)
