### 2021-11-17 #3917 @pdebuyl
@Saran-nns @mirca make sure to accept the invitation to the reviewers group and to have a look at the reviewer guidelines linked to at the top of this review page.

@shailesh1729 the article is very long by JOSS standard, I recommed to shorten it already. The explanations about l0 and l1 problems and about the linear operators are best left in the software documentation.
