### 2021-11-17 71886c9
moving theory stuff from paper to online docs
