+++ docs/ssc/index.rst
@@ -1,0 +1,8 @@
+ Sparse Subspace Clustering 
+ ==================================
+
+ .. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   intro
