+++ _sources/gallery/pursuit/cosamp_step_by_step.rst.txt
@@ -189,7 +189,7 @@ A sparse model vector
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7ff09a0d27c0>]
+    [<matplotlib.lines.Line2D object at 0x7f08a772d790>]
 
 
 
@@ -247,7 +247,7 @@ Compressive measurements
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7ff0995960a0>]
+    [<matplotlib.lines.Line2D object at 0x7f08a76b6130>]
 
 
 
@@ -627,7 +627,7 @@ We now have our first estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7ff099351130>
+    <matplotlib.legend.Legend object at 0x7f08a6dd2430>
 
 
 
@@ -1046,7 +1046,7 @@ Here is our updated estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7ff099010d90>
+    <matplotlib.legend.Legend object at 0x7f08a7115370>
 
 
 
@@ -1264,14 +1264,14 @@ Let's plot the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7ff099114dc0>
+    <matplotlib.legend.Legend object at 0x7f08a6f19940>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  5.629 seconds)
+   **Total running time of the script:** ( 0 minutes  6.594 seconds)
 
 
 .. _sphx_glr_download_gallery_pursuit_cosamp_step_by_step.py:
