+++ _sources/gallery/cluster/ssc_omp.rst.txt
@@ -305,7 +305,7 @@ Build the affinity matrix
  .. code-block:: none
 
 
-    <matplotlib.image.AxesImage object at 0x7ff09b27baf0>
+    <matplotlib.image.AxesImage object at 0x7f08a907e670>
 
 
 
@@ -503,7 +503,7 @@ Build the affinity matrix
  .. code-block:: none
 
 
-    <matplotlib.image.AxesImage object at 0x7ff09b0f7400>
+    <matplotlib.image.AxesImage object at 0x7f08a8edc610>
 
 
 
@@ -584,7 +584,7 @@ Evaluate the clustering performance
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  8.368 seconds)
+   **Total running time of the script:** ( 0 minutes  9.789 seconds)
 
 
 .. _sphx_glr_download_gallery_cluster_ssc_omp.py:
