+++ _sources/gallery/rec_l1/partial_wh_sensor_cosine_basis.rst.txt
@@ -113,7 +113,7 @@ Non-sparse signal
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7ff099407d90>]
+    [<matplotlib.lines.Line2D object at 0x7f08a7658a00>]
 
 
 
@@ -149,7 +149,7 @@ The Sparsifying Basis
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7ff0999f77c0>]
+    [<matplotlib.lines.Line2D object at 0x7f08a7c044f0>]
 
 
 
@@ -234,7 +234,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7ff099801af0>]
+    [<matplotlib.lines.Line2D object at 0x7f08a6e02460>]
 
 
 
@@ -307,14 +307,14 @@ Solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7ff098d591f0>
+    <matplotlib.legend.Legend object at 0x7f08a6bc2190>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  6.246 seconds)
+   **Total running time of the script:** ( 0 minutes  7.285 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_partial_wh_sensor_cosine_basis.py:
