+++ _sources/gallery/pursuit/cs1bit_biht.rst.txt
@@ -314,7 +314,7 @@ check if the support is recovered correctly
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.100 seconds)
+   **Total running time of the script:** ( 0 minutes  2.431 seconds)
 
 
 .. _sphx_glr_download_gallery_pursuit_cs1bit_biht.py:

[…]
