+++ _sources/source/_autosummary/cr.sparse.cluster.ssc.subspace_preservation_stats_jit.rst.txt
@@ -1,0 +1,6 @@
+﻿cr.sparse.cluster.ssc.subspace\_preservation\_stats\_jit
+========================================================
+
+.. currentmodule:: cr.sparse.cluster.ssc
+
+.. autofunction:: subspace_preservation_stats_jit
