+++ _sources/gallery/rec_l1/spikes_l1ls.rst.txt
@@ -123,7 +123,7 @@ The Spikes
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f08a6e0a3a0>]
+    [<matplotlib.lines.Line2D object at 0x7f9cea372d90>]
 
 
 
@@ -181,7 +181,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f08a77a2eb0>]
+    [<matplotlib.lines.Line2D object at 0x7f9ceae8b820>]
 
 
 
@@ -267,7 +267,7 @@ Solution
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f08a6a1d7f0>]
+    [<matplotlib.lines.Line2D object at 0x7f9ce9a78940>]
 
 
 
@@ -299,7 +299,7 @@ The magnitudes of non-zero values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f08a69f4c10>]
+    [<matplotlib.lines.Line2D object at 0x7f9ce99d9f40>]
 
 
 
@@ -335,7 +335,7 @@ Thresholding for large values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f08a691f2e0>]
+    [<matplotlib.lines.Line2D object at 0x7f9ce9978430>]
 
 
 
@@ -416,14 +416,14 @@ Improvement using least squares over support
     (1024, 160)
     ls_rel_error=2.057e-02
 
-    [<matplotlib.lines.Line2D object at 0x7f08a67d1430>]
+    [<matplotlib.lines.Line2D object at 0x7f9ce982e580>]
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  8.275 seconds)
+   **Total running time of the script:** ( 0 minutes  9.492 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_spikes_l1ls.py:
