+++ _sources/source/_autosummary/cr.sparse.cluster.ssc.angles_between_points.rst.txt
@@ -1,0 +1,6 @@
+﻿cr.sparse.cluster.ssc.angles\_between\_points
+=============================================
+
+.. currentmodule:: cr.sparse.cluster.ssc
+
+.. autofunction:: angles_between_points
