+++ _sources/gallery/sg_execution_times.rst.txt
@@ -5,8 +5,8 @@
 
 Computation times
 =================
-**00:01.963** total execution time for **gallery** files:
+**00:02.086** total execution time for **gallery** files:
 
 +---------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_gallery_sparse_vector_normals.py` (``sparse_vector_normals.py``) | 00:01.963 | 0.0 MB |
+| :ref:`sphx_glr_gallery_sparse_vector_normals.py` (``sparse_vector_normals.py``) | 00:02.086 | 0.0 MB |
 +---------------------------------------------------------------------------------+-----------+--------+
