+++ _sources/source/_autosummary/cr.sparse.cluster.ssc.batch_build_representation_omp.rst.txt
@@ -1,0 +1,6 @@
+﻿cr.sparse.cluster.ssc.batch\_build\_representation\_omp
+=======================================================
+
+.. currentmodule:: cr.sparse.cluster.ssc
+
+.. autofunction:: batch_build_representation_omp
