+++ _sources/source/_autosummary/cr.sparse.cluster.ssc.inn_positions.rst.txt
@@ -1,0 +1,6 @@
+﻿cr.sparse.cluster.ssc.inn\_positions
+====================================
+
+.. currentmodule:: cr.sparse.cluster.ssc
+
+.. autofunction:: inn_positions
