+++ _downloads/62389dad0fee8633e95d042ef09bde3c/ssc_omp.py
@@ -1,5 +1,5 @@
 """
-.. _gallery:cluster:ssc:omp
+.. _gallery:cluster:ssc:omp:
 
 Sparse Subspace Clustering - OMP
 =============================================
