+++ _sources/gallery/lop/wt_op.rst.txt
@@ -113,7 +113,7 @@ Sampling frequency
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9ceae8b430>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc117dd90>]
 
 
 
@@ -164,7 +164,7 @@ Wavelet coefficients
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9cea7dea00>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc104dca0>]
 
 
 
@@ -199,7 +199,7 @@ Let's keep only 10 percent of the coefficients
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9cea7b2fa0>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc0f84280>]
 
 
 
@@ -241,10 +241,10 @@ Reconstruction
 
  .. code-block:: none
 
-    0.03425936599788833
+    0.034259365997888334
     36.56352988951204
 
-    <matplotlib.legend.Legend object at 0x7f9cea694700>
+    <matplotlib.legend.Legend object at 0x7f3cc0eebfd0>
 
 
 
@@ -369,7 +369,7 @@ Reconstruction
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  10.807 seconds)
+   **Total running time of the script:** ( 0 minutes  8.411 seconds)
 
 
 .. _sphx_glr_download_gallery_lop_wt_op.py:
