+++ _sources/gallery/wavelets/frequency_change_detection.rst.txt
@@ -134,7 +134,7 @@ Test signal generation
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9ce99ddf70>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc0f7a220>]
 
 
 
@@ -175,7 +175,7 @@ Frequency spectrum
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9ce99367f0>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc0138a00>]
 
 
 
@@ -265,7 +265,7 @@ Second level detail coefficients
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9ce967d310>]
+    [<matplotlib.lines.Line2D object at 0x7f3cbfe29b50>]
 
 
 
@@ -313,7 +313,7 @@ correspond to t=12 sec.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.839 seconds)
+   **Total running time of the script:** ( 0 minutes  2.113 seconds)
 
 
 .. _sphx_glr_download_gallery_wavelets_frequency_change_detection.py:
