+++ _sources/gallery/rec_l1/partial_wh_sensor_cosine_basis.rst.txt
@@ -113,7 +113,7 @@ Non-sparse signal
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9cea7c0400>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc0631070>]
 
 
 
@@ -149,7 +149,7 @@ The Sparsifying Basis
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9cea639fd0>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc0c750a0>]
 
 
 
@@ -234,7 +234,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f9cea151100>]
+    [<matplotlib.lines.Line2D object at 0x7f3cc0ecf970>]
 
 
 
@@ -307,14 +307,14 @@ Solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f9ce9c102b0>
+    <matplotlib.legend.Legend object at 0x7f3cc0441bb0>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  8.353 seconds)
+   **Total running time of the script:** ( 0 minutes  6.218 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_partial_wh_sensor_cosine_basis.py:
