+++ _sources/source/_autosummary/cr.sparse.cluster.ssc.bcoo_to_sparse.rst.txt
@@ -1,0 +1,6 @@
+﻿cr.sparse.cluster.ssc.bcoo\_to\_sparse
+======================================
+
+.. currentmodule:: cr.sparse.cluster.ssc
+
+.. autofunction:: bcoo_to_sparse
