+++ _sources/gallery/wavelets/cameraman_decomposition.rst.txt
@@ -153,7 +153,7 @@ Plot the decomposition
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.988 seconds)
+   **Total running time of the script:** ( 0 minutes  0.716 seconds)
 
 
 .. _sphx_glr_download_gallery_wavelets_cameraman_decomposition.py:
