### 2021-11-17 #3917 @shailesh1729
@pdebuyl, I have revised the paper and moved the theoretical explanations to the online documentation and placed references to it in the paper draft.  It is down to 7 pages in draft.

Kindly, let me know if I should work on shortening it further.

With regards,
Shailesh
