### 2021-11-21 #3917 @Saran-nns
Hi, a quick update. I created an [issue](https://github.com/carnotresearch/cr-sparse/issues/21#issue-1059511222) with some suggestions at the repository.
