### 2021-11-22 #3917 @pdebuyl
Hi @shailesh1729 thanks for the editing, this is better. The listing of features in the summary is still excessive (explicit and detailed listing of all subpackages is also for the documentation). Also, a mention such as "For details, see the onlinedocumentation" is redundant.
