+++ paper/paper.md
@@ -35,7 +35,40 @@ standard signal processing transforms,
 and linear algebra subroutines for solving sparse linear systems. 
 It has been built using Google JAX [@jax2018github], which enables the same high level
 Python code to get efficiently compiled on CPU, GPU and TPU architectures
-using XLA.  
+using XLA [@abadi2017computational]. 
+
+![Sparse signal representations and compressive sensing](./srr_cs.png)
+
+Traditional signal processing exploits the underlying structure in signals
+by representing them using Fourier or wavelet orthonormal bases. 
+In these representations,
+most of the signal energy is concentrated in few coefficients allowing greater
+flexibility in analysis and processing of signals. More flexibility can be
+achieved by using overcomplete dictionaries [@mallat2008wavelet]
+(e.g. unions of orthonormal bases). However, the construction of
+sparse representations of signals in these overcomplete dictionaries 
+is no longer straightforward and requires use of specialized sparse
+coding algorithms like orthogonal matching pursuit [@pati1993orthogonal]
+or basis pursuit [@chen2001atomic]. The key idea behind these algorithms 
+is the fact that under-determined systems $A x = b$ can be solved efficiently
+to provide sparse solutions $x$ if the matrix $A$ satisfies specific conditions
+on its properties like coherence. Compressive sensing takes the same 
+idea in the other direction and contends that signals having sparse representations
+in suitable bases can be acquired by very few data-independent 
+random measurements $y = \Phi x$ if the sensing or measurement system $\Phi$
+satisfies certain conditions like restricted isometry property [@candes2008restricted].
+The same sparse coding algorithms can be tailored for sparse signal recovery
+from compressed measurements. 
+
+A short mathematical introduction to compressive sensing and sparse representation problems 
+is provided in [online documentation](https://cr-sparse.readthedocs.io/en/latest/intro.html).
+For comprehensive introduction to sparse
+representations and compressive sensing,
+ please refer to excellent books [@mallat2008wavelet;@elad2010sparse;@foucart2013mathintro],
+papers [@donoho2006compressed;@qaisar2013compressive],
+[Rice Compressive Sensing Resources](https://dsp.rice.edu/cs/) and references therein.
+
+# Package Overview
 
 The `cr.sparse.pursuit` package includes greedy and thresholding
 based solvers for sparse recovery. It includes: 
@@ -151,49 +184,27 @@ Several of these libraries contain key
 performance critical sub-routines
 in the form of C/C++ extensions making portability to GPUs harder. 
 
-There are some large libraries which focus on specific
-areas however they are generally CPU based. Additional effort 
-is spent in making the functionality available on GPUs.
+There are some Python libraries which focus on specific
+areas however they are generally CPU based.
+E.g., [`pyCSalgos`](https://github.com/nikcleju/pyCSalgos) is an
+old Python implementation of various Compressed Sensing algorithms.
+[`spgl1`](https://github.com/drrelyea/spgl1) is a `NumPy` based
+implementation of spectral projected gradient for L1 minimization.
+`c-lasso` [@simpson2021classo] is a Python package for constrained sparse regression 
+and classification. This is also CPU only. 
 [`PyWavelets`](https://github.com/PyWavelets/pywt) is an excellent 
 CPU only wavelets implementation in Python closely following the API
 of Wavelet toolbox in MATLAB. The performance critical parts have been
 written entirely in C. There are several attempts to port it on GPU
-using `PyTorch` or `Tensorflow` backends.
-[`PyLops`](https://github.com/PyLops/pylops) includes a GPU backend
-apart from its CPU implementation but the implementation appears to be
-somewhat complex.
-
[…]
