+++ paper/paper.bib
@@ -21,6 +21,20 @@
   publisher={IEEE}
 }
 
+@article{becker2011templates,
+  title={Templates for convex cone problems with applications to sparse signal recovery},
+  author={Becker, Stephen R and Cand{\`e}s, Emmanuel J and Grant, Michael C},
+  journal={Mathematical programming computation},
+  volume={3},
+  number={3},
+  pages={165},
+  year={2011},
+  doi={10.1007/s12532-011-0029-5},
+  url={https://doi.org/10.1007/s12532-011-0029-5},
+  publisher={Springer}
+}
+
+
 @article{blumensath2009iterative,
   title={Iterative hard thresholding for compressed sensing},
   author={Blumensath, Thomas and Davies, Mike E},
@@ -57,6 +71,32 @@
   doi={10.1109/wvm.1991.212809},
 	url={https://doi.org/10.1109/wvm.1991.212809},
   organization={IEEE}
+}
+
+@article{candes2008restricted,
+  title={The restricted isometry property and its implications for compressed sensing},
+  author={Candes, Emmanuel J},
+  journal={Comptes rendus mathematique},
+  volume={346},
+  number={9-10},
+  pages={589--592},
+  year={2008},
+  doi={10.1016/j.crma.2008.03.014},
+  url={https://doi.org/10.1016/j.crma.2008.03.014},
+  publisher={Elsevier}
+}
+
+@article{chen2001atomic,
+  title={Atomic decomposition by basis pursuit},
+  author={Chen, Scott Shaobing and Donoho, David L and Saunders, Michael A},
+  journal={SIAM review},
+  volume={43},
+  number={1},
+  pages={129--159},
+  year={2001},
+  doi={10.1137/S003614450037906X},
+  url={https://epubs.siam.org/doi/10.1137/S003614450037906X},
+  publisher={SIAM}
 }
 
 @article{costeira1998multibody,
@@ -98,6 +138,19 @@
   publisher={IEEE}
 }
 
+@article{donoho2006compressed,
+	doi = {10.1109/tit.2006.871582},
+	url = {https://doi.org/10.1109%2Ftit.2006.871582},
+	year = 2006,
+	month = {apr},
+	publisher = {Institute of Electrical and Electronics Engineers ({IEEE})},
+	volume = {52},
+	number = {4},
+	pages = {1289--1306},
+	author = {D.L. Donoho},
+	title = {Compressed sensing},
+	journal = {{IEEE} Transactions on Information Theory}
+}
 
 @article{dyer2013greedy,
   title={Greedy feature selection for subspace clustering},
@@ -153,6 +206,14 @@
   year={2011}
 }
 
+@book{foucart2013mathintro,
+	doi = {10.1007/978-0-8176-4948-7},
+	url = {https://doi.org/10.1007%2F978-0-8176-4948-7},
+	year = 2013,
+	publisher = {Springer New York},
+	author = {Simon Foucart and Holger Rauhut},
+	title = {A Mathematical Introduction to Compressive Sensing}
+}
 
 @article{gear1998multibody,
   title={Multibody grouping from motion images},
@@ -263,6 +324,19 @@ address = {Natick, Massachusetts}
   publisher={Trelgol Publishing USA}
 }
 
+@article{parikh2014proximal,
+  title={Proximal algorithms},
+  author={Parikh, Neal and Boyd, Stephen},
+  journal={Foundations and Trends in optimization},
+  volume={1},
+  number={3},
+  pages={127--239},
+  year={2014},
+  url = {https://doi.org/10.1561/2400000003},
+  doi = {10.1561/2400000003},
+  publisher={Now Publishers Inc. Hanover, MA, USA}
+}
+
 @inproceedings{pati1993orthogonal,
[…]
