+++ docs/intro.rst
@@ -1,7 +1,7 @@
 Introduction
 =====================
 
-.. contents::
+.. contents::`
     :depth: 2
     :local:
 

[…]
