+++ _sources/gallery/cluster/ssc_omp.rst.txt
@@ -307,7 +307,7 @@ Build the affinity matrix
  .. code-block:: none
 
 
-    <matplotlib.image.AxesImage object at 0x7f3cc2901d60>
+    <matplotlib.image.AxesImage object at 0x7f7285668610>
 
 
 
@@ -505,7 +505,7 @@ Build the affinity matrix
  .. code-block:: none
 
 
-    <matplotlib.image.AxesImage object at 0x7f3cc276eb20>
+    <matplotlib.image.AxesImage object at 0x7f72854ba610>
 
 
 
@@ -586,7 +586,7 @@ Evaluate the clustering performance
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  8.422 seconds)
+   **Total running time of the script:** ( 0 minutes  9.308 seconds)
 
 
 .. _sphx_glr_download_gallery_cluster_ssc_omp.py:
