+++ _sources/gallery/pursuit/cosamp_step_by_step.rst.txt
@@ -189,7 +189,7 @@ A sparse model vector
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f3cc117d340>]
+    [<matplotlib.lines.Line2D object at 0x7f728457e100>]
 
 
 
@@ -247,7 +247,7 @@ Compressive measurements
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f3cc0c85100>]
+    [<matplotlib.lines.Line2D object at 0x7f7283a90310>]
 
 
 
@@ -627,7 +627,7 @@ We now have our first estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f3cc0f9b8e0>
+    <matplotlib.legend.Legend object at 0x7f7283cf0220>
 
 
 
@@ -1046,7 +1046,7 @@ Here is our updated estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f3cc0fd27f0>
+    <matplotlib.legend.Legend object at 0x7f7283c7e160>
 
 
 
@@ -1264,14 +1264,14 @@ Let's plot the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f3cc09cfe20>
+    <matplotlib.legend.Legend object at 0x7f7283578790>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  5.566 seconds)
+   **Total running time of the script:** ( 0 minutes  6.146 seconds)
 
 
 .. _sphx_glr_download_gallery_pursuit_cosamp_step_by_step.py:
