+++ paper/paper.bib
@@ -294,6 +294,18 @@
   publisher={Elsevier}
 }
 
+@article{marques2018review,
+  title={A review of sparse recovery algorithms},
+  author={Marques, Elaine Crespo and Maciel, Nilson and Naviner, Lirida and Cai, Hao and Yang, Jun},
+  journal={IEEE access},
+  volume={7},
+  pages={1300--1322},
+  year={2018},
+  doi={10.1109/ACCESS.2018.2886471},
+  url={https://doi.org/10.1109/ACCESS.2018.2886471},
+  publisher={IEEE}
+}
+
 @book{MATLAB:2018,
 year = {2018},
 author = {MATLAB},
