+++ _sources/gallery/sg_execution_times.rst.txt
@@ -5,8 +5,8 @@
 
 Computation times
 =================
-**00:01.864** total execution time for **gallery** files:
+**00:02.140** total execution time for **gallery** files:
 
 +---------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_gallery_sparse_vector_normals.py` (``sparse_vector_normals.py``) | 00:01.864 | 0.0 MB |
+| :ref:`sphx_glr_gallery_sparse_vector_normals.py` (``sparse_vector_normals.py``) | 00:02.140 | 0.0 MB |
 +---------------------------------------------------------------------------------+-----------+--------+
