+++ _sources/requirements.txt
@@ -1,0 +1,24 @@
+chex>=0.0.4
+jax>=0.1.55
+jaxlib>=0.1.37
+numpy>=1.18.0
+sympy>=1.6
+scipy
+matplotlib
+imageio
+click
+requests
+pillow
+PyWavelets
+scikit-image
+scikit-learn
+sphinx==4.0.0
+sphinxcontrib-katex>=0.7.1
+sphinxcontrib-bibtex>=1.0.0
+sphinx-autodoc-typehints>=1.11.1
+IPython>=7.16.1
+ipykernel>=5.3.4
+nbsphinx>=0.8.0
+sphinx-gallery>=0.8.0
+sphinx-panels
+myst-parser
