+++ _sources/gallery/pursuit/cosamp_step_by_step.rst.txt
@@ -155,7 +155,7 @@ Coherence of atoms in the sensing matrix
 
  .. code-block:: none
 
-    0.3881940752728345
+    0.38819407527283445
 
 
 
@@ -189,7 +189,7 @@ A sparse model vector
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f728457e100>]
+    [<matplotlib.lines.Line2D object at 0x7f87de0f3fd0>]
 
 
 
@@ -247,7 +247,7 @@ Compressive measurements
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f7283a90310>]
+    [<matplotlib.lines.Line2D object at 0x7f87ddf7a220>]
 
 
 
@@ -317,7 +317,7 @@ Squared norm/energy of the residual
 
  .. code-block:: none
 
-    r_norm_sqr=7.40121202914161
+    r_norm_sqr=7.401212029141609
 
 
 
@@ -627,7 +627,7 @@ We now have our first estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f7283cf0220>
+    <matplotlib.legend.Legend object at 0x7f87dd8f69d0>
 
 
 
@@ -1046,7 +1046,7 @@ Here is our updated estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f7283c7e160>
+    <matplotlib.legend.Legend object at 0x7f87ddc148e0>
 
 
 
@@ -1089,7 +1089,7 @@ Compute the residual and verify that it is now below the allowed tolerance
 
  .. code-block:: none
 
-    r_norm_sqr=1.76e-30 < max_r_norm_sqr=7.40e-06
+    r_norm_sqr=3.05e-30 < max_r_norm_sqr=7.40e-06
 
 
 
@@ -1230,7 +1230,7 @@ Print the residual energy and the number of iterations when the algorithm conver
 
  .. code-block:: none
 
-    1.0297848394871909e-29 3
+    2.14594481109173e-29 3
 
 
 
@@ -1264,14 +1264,14 @@ Let's plot the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f7283578790>
+    <matplotlib.legend.Legend object at 0x7f87ddd36f10>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  6.146 seconds)
+   **Total running time of the script:** ( 0 minutes  7.151 seconds)
 
 
 .. _sphx_glr_download_gallery_pursuit_cosamp_step_by_step.py:
