### 2021-11-22 carnotresearch/cr-sparse#21 @shailesh1729
Dear @Saran-nns,

Thank you very much for the detailed feedback. I have updated the paper draft in openjournals/joss-reviews#3917.  I am also trying to balance between the paper length concerns mentioned by pdebuyl and adding more content to address the concerns listed here.

Some specific answers:

- I have added a very short intro to sparse problems. I have kept the more detailed mathematical intro in the [online docs](https://cr-sparse.readthedocs.io/en/latest/intro.html).  Also have provided references to relevant books and papers.
- Have removed JAX description from the statement of need section. Moved some of the discussion into the new limitations section as relevant.
- See `here <https://cr-sparse.readthedocs.io/en/latest/algorithms.html>`_ in the docs for the list of algorithms supported (and planned) in organized hierarchically. I have added a reference to this page in README.  I will slowly enrich this page further with more information later as I build more detailed usage and application docs for individual algorithms.
