### 2021-11-22 carnotresearch/cr-sparse#21 @Saran-nns
Thanks for working on my suggestions. Besides the paper length, I am happy with the paper draft and the documentation. I close this issue in your favor :)
