### 2021-11-23 e7e59cc
some acronyms added
