+++ _sources/gallery/rec_l1/partial_wh_sensor_cosine_basis.rst.txt
@@ -113,7 +113,7 @@ Non-sparse signal
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f87ddc53d30>]
+    [<matplotlib.lines.Line2D object at 0x7f44b9bffcd0>]
 
 
 
@@ -149,7 +149,7 @@ The Sparsifying Basis
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f87de7eed30>]
+    [<matplotlib.lines.Line2D object at 0x7f44b93e04c0>]
 
 
 
@@ -234,7 +234,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f87dd905940>]
+    [<matplotlib.lines.Line2D object at 0x7f44b974d370>]
 
 
 
@@ -307,14 +307,14 @@ Solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f87dd647970>
+    <matplotlib.legend.Legend object at 0x7f44b90c1d60>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  8.617 seconds)
+   **Total running time of the script:** ( 0 minutes  8.171 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_partial_wh_sensor_cosine_basis.py:
