### 2021-11-23 #3917 @Saran-nns
@shailesh1729 Besides the length of the draft as mentioned by @pdebuyl, I am convinced with the recent draft. I closed the issue https://github.com/carnotresearch/cr-sparse/issues/21 in favor of @shailesh1729
