### 2021-11-23 carnotresearch/cr-sparse#22 @Saran-nns
[Joss Review] Readme Update
Hi @shailesh1729

I found the readme is somewhat cluttered.

Here are a few of my suggestions;

- [x] Quickstart section is not necessary for this readme. I recommend removing it

- [x] Please include the image about the sparse representation problem(at draft summary) below lines[14-25].

- [x] Lines[27-31]  can be replaced by _For detailed documentation and usage_, please visit https://cr-sparse.readthedocs.io/en/latest.

- [x] Lines [34-51] looks redundant. I suggest removing it since you have mentioned the algorithms already in the introduction part and are present in the _Examples_ section also. For more examples please direct the users to the example gallery https://cr-sparse.readthedocs.io/en/latest/gallery/index.html

- [x] Line 57 is redundant. Remove line 57 and move line 56 to line 73

Thank you and let me know what you think about the suggestions,
