+++ _sources/gallery/lop/deconvolution.rst.txt
@@ -370,7 +370,7 @@ We run LSQR algorithm to clean the signal contents for 100 iterations
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.999 seconds)
+   **Total running time of the script:** ( 0 minutes  3.273 seconds)
 
 
 .. _sphx_glr_download_gallery_lop_deconvolution.py:
