+++ _sources/gallery/sparse_vector_normals.rst.txt
@@ -174,7 +174,7 @@ Let's plot the vector to see where the non-zero entries are
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  1.969 seconds)
+   **Total running time of the script:** ( 0 minutes  2.198 seconds)
 
 
 .. _sphx_glr_download_gallery_sparse_vector_normals.py:
