+++ _sources/gallery/lop/wt_op.rst.txt
@@ -113,7 +113,7 @@ Sampling frequency
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f44b9f56ac0>]
+    [<matplotlib.lines.Line2D object at 0x7f608e858a00>]
 
 
 
@@ -164,7 +164,7 @@ Wavelet coefficients
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f44b9d28b50>]
+    [<matplotlib.lines.Line2D object at 0x7f608e6c1fa0>]
 
 
 
@@ -199,7 +199,7 @@ Let's keep only 10 percent of the coefficients
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f44b9d08130>]
+    [<matplotlib.lines.Line2D object at 0x7f608e6ab640>]
 
 
 
@@ -244,7 +244,7 @@ Reconstruction
     0.03425936599788833
     36.56352988951204
 
-    <matplotlib.legend.Legend object at 0x7f44b9be8250>
+    <matplotlib.legend.Legend object at 0x7f608e58b580>
 
 
 
@@ -369,7 +369,7 @@ Reconstruction
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  10.487 seconds)
+   **Total running time of the script:** ( 0 minutes  11.069 seconds)
 
 
 .. _sphx_glr_download_gallery_lop_wt_op.py:
