+++ _sources/gallery/cluster/kmeans.rst.txt
@@ -114,7 +114,7 @@ Prepare sample data
  .. code-block:: none
 
 
-    <matplotlib.collections.PathCollection object at 0x7f44d8e547f0>
+    <matplotlib.collections.PathCollection object at 0x7f6099ef46a0>
 
 
 
@@ -157,7 +157,7 @@ K-means clustering
  .. code-block:: none
 
 
-    <matplotlib.collections.PathCollection object at 0x7f44d8d5d730>
+    <matplotlib.collections.PathCollection object at 0x7f6097c9f5e0>
 
 
 
@@ -294,7 +294,7 @@ For any new point, it is easy to find the index of nearest centroid
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  5.260 seconds)
+   **Total running time of the script:** ( 0 minutes  5.643 seconds)
 
 
 .. _sphx_glr_download_gallery_cluster_kmeans.py:
