+++ _sources/gallery/wavelets/chirp_cwt_mexh.rst.txt
@@ -229,7 +229,7 @@ voices per octave
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  6.005 seconds)
+   **Total running time of the script:** ( 0 minutes  6.379 seconds)
 
 
 .. _sphx_glr_download_gallery_wavelets_chirp_cwt_mexh.py:
