+++ docs/start.rst
@@ -34,10 +34,10 @@ Directly from our GITHUB repository:
 
 Examples
 ----------------
-See the :ref:`examples gallery <gallery>`.
 
-A more extensive collection of example notebooks is available in the `companion repository <https://github.com/carnotresearch/cr-sparse-companion>`_.
-Some micro-benchmarks are reported `here <https://github.com/carnotresearch/cr-sparse/blob/master/paper/paper.md#runtime-comparisons>`_.
+* See the :ref:`examples gallery <gallery>`.
+* A more extensive collection of example notebooks is available in the `companion repository <https://github.com/carnotresearch/cr-sparse-companion>`_.
+* Some micro-benchmarks are reported `here <https://github.com/carnotresearch/cr-sparse/blob/master/paper/paper.md#runtime-comparisons>`_.
 
 
 
