+++ _sources/gallery/cluster/ssc_omp.rst.txt
@@ -159,7 +159,7 @@ Print the minimum angle between any pair of subspaces
 
  .. code-block:: none
 
-    53.35521088678854
+    53.35521088678853
 
 
 
@@ -307,7 +307,7 @@ Build the affinity matrix
  .. code-block:: none
 
 
-    <matplotlib.image.AxesImage object at 0x7f608ff1d8b0>
+    <matplotlib.image.AxesImage object at 0x7f40c4a74850>
 
 
 
@@ -505,7 +505,7 @@ Build the affinity matrix
  .. code-block:: none
 
 
-    <matplotlib.image.AxesImage object at 0x7f608fd7e850>
+    <matplotlib.image.AxesImage object at 0x7f40c48c9850>
 
 
 
@@ -547,13 +547,13 @@ Predicted cluster labels
 
  .. code-block:: none
 
-    [3 1 4 4 2 0 3 3 1 1 3 1 0 4 0 1 1 3 4 2 1 0 2 1 3 4 0 2 3 0 4 4 1 3 4 0 1
-     1 3 3 2 3 0 1 2 2 0 0 2 2 4 0 2 3 1 0 2 0 3 2 2 4 3 2 1 3 1 0 3 3 0 2 2 2
-     1 4 0 3 4 1 1 2 4 1 4 2 2 0 1 0 4 3 3 0 0 2 0 3 4 4 3 0 0 4 3 2 2 2 0 2 4
-     4 0 4 3 3 0 4 2 3 4 1 1 1 2 1 0 1 3 0 1 0 4 2 1 1 0 1 0 1 2 0 0 4 3 3 2 2
-     2 2 2 1 1 0 4 2 4 2 0 0 1 4 1 3 3 2 0 2 3 0 1 2 2 4 1 2 4 3 3 4 0 4 4 4 4
-     3 1 1 2 3 1 4 3 0 3 3 0 0 3 2 3 2 2 0 1 1 2 1 4 4 1 3 0 4 0 1 4 1 3 1 4 1
-     1 3 0 4 3 4 4 0 3 4 2 4 2 3 1 3 1 4 2 3 4 2 0 0 4 3 4 0]
+    [3 0 1 1 2 4 3 3 0 0 3 0 4 1 4 0 0 3 1 2 0 4 2 0 3 1 4 2 3 4 1 1 0 3 1 4 0
+     0 3 3 2 3 4 0 2 2 4 4 2 2 1 4 2 3 0 4 2 4 3 2 2 1 3 2 0 3 0 4 3 3 4 2 2 2
+     0 1 4 3 1 0 0 2 1 0 1 2 2 4 0 4 1 3 3 4 4 2 4 3 1 1 3 4 4 1 3 2 2 2 4 2 1
+     1 4 1 3 3 4 1 2 3 1 0 0 0 2 0 4 0 3 4 0 4 1 2 0 0 4 0 4 0 2 4 4 1 3 3 2 2
+     2 2 2 0 0 4 1 2 1 2 4 4 0 1 0 3 3 2 4 2 3 4 0 2 2 1 0 2 1 3 3 1 4 1 1 1 1
+     3 0 0 2 3 0 1 3 4 3 3 4 4 3 2 3 2 2 4 0 0 2 0 1 1 0 3 4 1 4 0 1 0 3 0 1 0
+     0 3 4 1 3 1 1 4 3 1 2 1 2 3 0 3 0 1 2 3 1 2 4 4 1 3 1 4]
 
 
 
@@ -586,7 +586,7 @@ Evaluate the clustering performance
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  11.332 seconds)
[…]
