+++ _sources/gallery/lop/matrix_operator.rst.txt
@@ -249,7 +249,7 @@ Compute A^H x
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.091 seconds)
+   **Total running time of the script:** ( 0 minutes  0.066 seconds)
 
 
 .. _sphx_glr_download_gallery_lop_matrix_operator.py:

[…]
