+++ _sources/gallery/rec_l1/spikes_l1ls.rst.txt
@@ -123,7 +123,7 @@ The Spikes
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f40c2b24940>]
+    [<matplotlib.lines.Line2D object at 0x7f5e477f47c0>]
 
 
 
@@ -181,7 +181,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f40c295b580>]
+    [<matplotlib.lines.Line2D object at 0x7f5e477f90a0>]
 
 
 
@@ -267,7 +267,7 @@ Solution
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f40c244f910>]
+    [<matplotlib.lines.Line2D object at 0x7f5e4740f520>]
 
 
 
@@ -299,7 +299,7 @@ The magnitudes of non-zero values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f40c2430f70>]
+    [<matplotlib.lines.Line2D object at 0x7f5e473e4940>]
 
 
 
@@ -335,7 +335,7 @@ Thresholding for large values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f40c234f400>]
+    [<matplotlib.lines.Line2D object at 0x7f5e472fffd0>]
 
 
 
@@ -416,14 +416,14 @@ Improvement using least squares over support
     (1024, 160)
     ls_rel_error=2.057e-02
 
-    [<matplotlib.lines.Line2D object at 0x7f40c227e550>]
+    [<matplotlib.lines.Line2D object at 0x7f5e47241160>]
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  7.211 seconds)
+   **Total running time of the script:** ( 0 minutes  7.621 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_spikes_l1ls.py:
