+++ _sources/gallery/pursuit/cosamp_step_by_step.rst.txt
@@ -189,7 +189,7 @@ A sparse model vector
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f40c31b1790>]
+    [<matplotlib.lines.Line2D object at 0x7f5e488e2940>]
 
 
 
@@ -247,7 +247,7 @@ Compressive measurements
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f40c2d6ccd0>]
+    [<matplotlib.lines.Line2D object at 0x7f5e47e69190>]
 
 
 
@@ -627,7 +627,7 @@ We now have our first estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f40c27f2f10>
+    <matplotlib.legend.Legend object at 0x7f5e477e4a60>
 
 
 
@@ -1046,7 +1046,7 @@ Here is our updated estimate of the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f40c3198e50>
+    <matplotlib.legend.Legend object at 0x7f5e481b29a0>
 
 
 
@@ -1264,14 +1264,14 @@ Let's plot the solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f40c29804f0>
+    <matplotlib.legend.Legend object at 0x7f5e47b54fa0>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  5.490 seconds)
+   **Total running time of the script:** ( 0 minutes  6.001 seconds)
 
 
 .. _sphx_glr_download_gallery_pursuit_cosamp_step_by_step.py:
