### 2021-11-24 carnotresearch/cr-sparse#22 @shailesh1729
Thanks, @Saran-nns  for the detailed feedback. A good look by a second pair of eyes really helps in cleaning up things.

I had done a hack earlier where I was including the `README.rst` also as a quickstart page in the online docs so that I could reuse the text.  This hack was also causing problems in including the image in the README. I have separated the quick start page in online docs from the main README now.

Hope that the new version of README addresses your concerns and is looking cleaner.
