+++ _sources/gallery/rec_l1/partial_wh_sensor_cosine_basis.rst.txt
@@ -113,7 +113,7 @@ Non-sparse signal
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e47f188b0>]
+    [<matplotlib.lines.Line2D object at 0x7fac36b4d910>]
 
 
 
@@ -149,7 +149,7 @@ The Sparsifying Basis
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e48702370>]
+    [<matplotlib.lines.Line2D object at 0x7fac373e7ee0>]
 
 
 
@@ -234,7 +234,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e477dae80>]
+    [<matplotlib.lines.Line2D object at 0x7fac36ee8100>]
 
 
 
@@ -307,14 +307,14 @@ Solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7f5e47515e20>
+    <matplotlib.legend.Legend object at 0x7fac368bc340>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  6.752 seconds)
+   **Total running time of the script:** ( 0 minutes  8.396 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_partial_wh_sensor_cosine_basis.py:
