+++ _sources/gallery/rec_l1/spikes_l1ls.rst.txt
@@ -123,7 +123,7 @@ The Spikes
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e477f47c0>]
+    [<matplotlib.lines.Line2D object at 0x7fac386b63a0>]
 
 
 
@@ -181,7 +181,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e477f90a0>]
+    [<matplotlib.lines.Line2D object at 0x7fac37394400>]
 
 
 
@@ -267,7 +267,7 @@ Solution
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e4740f520>]
+    [<matplotlib.lines.Line2D object at 0x7fac367b7970>]
 
 
 
@@ -299,7 +299,7 @@ The magnitudes of non-zero values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e473e4940>]
+    [<matplotlib.lines.Line2D object at 0x7fac36799fa0>]
 
 
 
@@ -335,7 +335,7 @@ Thresholding for large values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7f5e472fffd0>]
+    [<matplotlib.lines.Line2D object at 0x7fac366b4460>]
 
 
 
@@ -416,14 +416,14 @@ Improvement using least squares over support
     (1024, 160)
     ls_rel_error=2.057e-02
 
-    [<matplotlib.lines.Line2D object at 0x7f5e47241160>]
+    [<matplotlib.lines.Line2D object at 0x7fac365ed5b0>]
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  7.621 seconds)
+   **Total running time of the script:** ( 0 minutes  9.147 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_spikes_l1ls.py:
