+++ _sources/gallery/cluster/kmeans.rst.txt
@@ -114,7 +114,7 @@ Prepare sample data
  .. code-block:: none
 
 
-    <matplotlib.collections.PathCollection object at 0x7f5e683505e0>
+    <matplotlib.collections.PathCollection object at 0x7fac607a8670>
 
 
 
@@ -157,7 +157,7 @@ K-means clustering
  .. code-block:: none
 
 
-    <matplotlib.collections.PathCollection object at 0x7f5e6822d520>
+    <matplotlib.collections.PathCollection object at 0x7fac6055f5b0>
 
 
 
@@ -294,7 +294,7 @@ For any new point, it is easy to find the index of nearest centroid
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  4.607 seconds)
+   **Total running time of the script:** ( 0 minutes  5.585 seconds)
 
 
 .. _sphx_glr_download_gallery_cluster_kmeans.py:
