### 2021-11-24 529353d
missing pandas dependency for docs [readthedocs]
