+++ docs/requirements.txt
@@ -9,6 +9,7 @@ imageio
 click
 requests
 pillow
+pandas
 PyWavelets
 scikit-image
 scikit-learn
