+++ _sources/gallery/rec_l1/spikes_l1ls.rst.txt
@@ -123,7 +123,7 @@ The Spikes
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7fac386b63a0>]
+    [<matplotlib.lines.Line2D object at 0x7fbbaca31790>]
 
 
 
@@ -181,7 +181,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7fac37394400>]
+    [<matplotlib.lines.Line2D object at 0x7fbbac182430>]
 
 
 
@@ -267,7 +267,7 @@ Solution
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7fac367b7970>]
+    [<matplotlib.lines.Line2D object at 0x7fbbabace8b0>]
 
 
 
@@ -299,7 +299,7 @@ The magnitudes of non-zero values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7fac36799fa0>]
+    [<matplotlib.lines.Line2D object at 0x7fbbabaa4cd0>]
 
 
 
@@ -335,7 +335,7 @@ Thresholding for large values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7fac366b4460>]
+    [<matplotlib.lines.Line2D object at 0x7fbbab9cd3a0>]
 
 
 
@@ -416,14 +416,14 @@ Improvement using least squares over support
     (1024, 160)
     ls_rel_error=2.057e-02
 
-    [<matplotlib.lines.Line2D object at 0x7fac365ed5b0>]
+    [<matplotlib.lines.Line2D object at 0x7fbbab8ff4f0>]
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  9.147 seconds)
+   **Total running time of the script:** ( 0 minutes  7.922 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_spikes_l1ls.py:
