+++ _sources/gallery/pursuit/cs1bit_biht.rst.txt
@@ -110,7 +110,7 @@ Sensing Matrix
 
  .. code-block:: none
 
-    2.400341648379146
+    2.4003416483791473
 
     Text(0.5, 1.0, '$\\Phi$')
 
@@ -306,7 +306,7 @@ check if the support is recovered correctly
  .. code-block:: none
 
     True
-    0.024352999672904285
+    0.02435299967290439
 
 
 
@@ -314,7 +314,7 @@ check if the support is recovered correctly
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.686 seconds)
+   **Total running time of the script:** ( 0 minutes  2.334 seconds)
 
 
 .. _sphx_glr_download_gallery_pursuit_cs1bit_biht.py:
