### 2021-11-24 carnotresearch/cr-sparse#22 @Saran-nns
Hi @shailesh1729 Looks much better now. Few contents from Quickstart section is still present, lines 8-10.

Also, I prefer the below order in Readme,
1. Introduction
2. Supported Platforms
3. Installation
4. Examples/Usage
5. **Contribution Guidelines/Code of Conduct** You can just provide the link to _Code of Conduct_ that you already have in the repository
6. Citation

Let me know what you think about the suggestions
