+++ _sources/gallery/lop/deblurring.rst.txt
@@ -379,7 +379,7 @@ We combine the convolution operator and the wavelet basis operator
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  24.632 seconds)
+   **Total running time of the script:** ( 0 minutes  25.077 seconds)
 
 
 .. _sphx_glr_download_gallery_lop_deblurring.py:
