+++ _sources/gallery/cluster/kmeans.rst.txt
@@ -114,7 +114,7 @@ Prepare sample data
  .. code-block:: none
 
 
-    <matplotlib.collections.PathCollection object at 0x7fbbccb25460>
+    <matplotlib.collections.PathCollection object at 0x7fab105165e0>
 
 
 
@@ -157,7 +157,7 @@ K-means clustering
  .. code-block:: none
 
 
-    <matplotlib.collections.PathCollection object at 0x7fbbcca1c3a0>
+    <matplotlib.collections.PathCollection object at 0x7fab1041d520>
 
 
 
@@ -294,7 +294,7 @@ For any new point, it is easy to find the index of nearest centroid
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  4.813 seconds)
+   **Total running time of the script:** ( 0 minutes  5.096 seconds)
 
 
 .. _sphx_glr_download_gallery_cluster_kmeans.py:
