+++ paper/paper.bib
@@ -479,6 +479,19 @@ address = {Natick, Massachusetts}
   publisher={IEEE}
 }
 
+@article{yang2011alternating,
+  title={Alternating direction algorithms for l_1-problems in compressive sensing},
+  author={Yang, Junfeng and Zhang, Yin},
+  journal={SIAM journal on scientific computing},
+  volume={33},
+  number={1},
+  pages={250--278},
+  year={2011},
+  doi={10.1137/090777761},
+  url={https://doi.org/10.1137/090777761},
+  publisher={SIAM}
+}
+
 @article{you2015sparse,
   title={Sparse subspace clustering by orthogonal matching pursuit},
   author={You, Chong and Vidal, Ren{\'e}},
