### 2021-11-28 54cff0a
Introducing ProxCapable wrapper
