### 2021-11-28 58bc4cd
introduce SmoothFunction wrapper
