+++ src/cr/sparse/opt/smooth.py
@@ -16,7 +16,13 @@
 from cr.sparse._src.opt.smooth import (
     smooth_value_grad
 )
+from cr.sparse._src.opt.smooth.smooth import (
+    SmoothFunction
+)
 
+from cr.sparse._src.opt.smooth.smooth import build as smooth_build
+from cr.sparse._src.opt.smooth.smooth import build2 as smooth_build2
+from cr.sparse._src.opt.smooth.smooth import build3 as smooth_build3
 
 from cr.sparse._src.opt.smooth.constant import smooth_constant
 from cr.sparse._src.opt.smooth.entropy import (
