### 2021-11-28 148ccac
smooth_func_translate
