### 2021-11-28 45f8a97
arr_rnorm_sqr, arr_rnorm
