### 2021-11-28 e5b4117
first working implementation of FOCS
