+++ src/cr/sparse/focs.py
@@ -22,3 +22,7 @@ from cr.sparse._src.focs.util import (
 from cr.sparse._src.focs.focs import focs
 from cr.sparse._src.focs.l1rls import l1rls
 
+from cr.sparse._src.focs.defs import (
+    FOCSOptions,
+    FOCSState
+)
