+++ src/cr/sparse/_src/focs/focs_at.py
@@ -16,15 +16,6 @@ from .defs import FOCSOptions, FOCSState
 
 import jax.numpy as jnp
 
-def focs_init(f, h, A, b, x0, options: FOCSOptions):
-    """
-    """
-    in_shape = A.input_shape
-    out_shape = A.output_shape
-    x = x0
-    f_v = jnp.inf
-    state = FOCSState(L=options.L0, theta=jnp.inf)
-    return state,
 
 def focs_at(f, h, A, b, x0, options: FOCSOptions):
     """First order conic solver Auslender and Teboulle method
