### 2021-11-28 40a3970
make it jit friendly
