+++ paper/paper.bib
@@ -21,6 +21,19 @@
   publisher={IEEE}
 }
 
+@article{beck2009fast,
+  title={A fast iterative shrinkage-thresholding algorithm for linear inverse problems},
+  author={Beck, Amir and Teboulle, Marc},
+  journal={SIAM journal on imaging sciences},
+  volume={2},
+  number={1},
+  pages={183--202},
+  year={2009},
+  doi={10.1137/080716542},
+  url={https://doi.org/10.1137/080716542},
+  publisher={SIAM}
+}
+
 @article{becker2011templates,
   title={Templates for convex cone problems with applications to sparse signal recovery},
   author={Becker, Stephen R and Cand{\`e}s, Emmanuel J and Grant, Michael C},
@@ -123,6 +136,19 @@
   doi={10.1109/tit.2009.2016006},
 	url={https://doi.org/10.1109/tit.2009.2016006},
   publisher={IEEE}
+}
+
+@article{daubechies2004iterative,
+  title={An iterative thresholding algorithm for linear inverse problems with a sparsity constraint},
+  author={Daubechies, Ingrid and Defrise, Michel and De Mol, Christine},
+  journal={Communications on Pure and Applied Mathematics: A Journal Issued by the Courant Institute of Mathematical Sciences},
+  volume={57},
+  number={11},
+  pages={1413--1457},
+  year={2004},
+  doi={10.1002/cpa.20042},
+  url={https://doi.org/10.1002/cpa.20042},
+  publisher={Wiley Online Library}
 }
 
 @article{davenport2010analysis,
@@ -259,6 +285,19 @@
   publisher={{IEEE} Comput. Soc}
 }
 
+@article{kim2007interior,
+  title={An interior-point method for large-scale $$\backslash$ell\_1 $-regularized least squares},
+  author={Kim, Seung-Jean and Koh, Kwangmoo and Lustig, Michael and Boyd, Stephen and Gorinevsky, Dimitry},
+  journal={IEEE journal of selected topics in signal processing},
+  volume={1},
+  number={4},
+  pages={606--617},
+  year={2007},
+  doi={10.1109/JSTSP.2007.910971},
+  url={https://doi.org/10.1109/JSTSP.2007.910971},
+  publisher={IEEE}
+}
+
 @article{lee2005acquiring,
   title={Acquiring linear subspaces for face recognition under variable lighting},
   author={Lee, Kuang-Chih and Ho, Jeffrey and Kriegman, David},
@@ -334,6 +373,17 @@ address = {Natick, Massachusetts}
   volume={1},
   year={2006},
   publisher={Trelgol Publishing USA}
+}
+
+@article{paige1982lsqr,
+  title={LSQR: An algorithm for sparse linear equations and sparse least squares},
+  author={Paige, Christopher C and Saunders, Michael A},
+  journal={ACM Transactions on Mathematical Software (TOMS)},
+  volume={8},
+  number={1},
+  pages={43--71},
+  year={1982},
+  publisher={ACM New York, NY, USA}
 }
 
 @article{parikh2014proximal,
