+++ paper/paper.md
@@ -18,7 +18,7 @@ affiliations:
       index: 1
 
 
-date: 5 November 2021
+date: 29 November 2021
 bibliography: paper.bib
 ---
 
@@ -72,56 +72,21 @@ papers [@donoho2006compressed;@qaisar2013compressive;@marques2018review],
 
 The `cr.sparse.pursuit` package includes greedy and thresholding
 based solvers for sparse recovery. It includes: 
-`OMP`, `CoSaMP`, `HTP`, `IHT`, `SP` algorithms. Normalized versions
-of `HTP` and `IHT` algorithms are also supported.
-These algorithms can work with unstructured random and dense sensing matrices
-as well as structured sensing matrices represented as linear operators
+`OMP`, `CoSaMP`, `HTP`, `IHT`, `SP` algorithms.
 (provided in `cr.sparse.lop` package).
-
-The `cr.sparse.cvx.adm.yall1` package includes a JAX version of 
-`Your algorithms for l1` [@yang2011alternating;@zhang2009user], an efficient solver
-for l1-minimization problems using 
-ADMM (alternating direction method of multipliers) approach.
-The `cr.sparse.cvx.l1ls` includes 
-*Truncated Newton Interior Points Method* for solving 
-the l1-minimization problem. 
-
-
+The `cr.sparse.cvx` package includes efficient solvers
+for l1-minimization problems using convex optimization methods.
 The `cr.sparse.sls` package provides JAX versions of
 `LSQR`, `ISTA`, `FISTA`  algorithms for solving sparse linear 
-systems. The linear system is represented by a linear operator
-from `cr.sparse.lop`. 
-
-The `cr.sparse.dict` package provides matrix versions of a variety
-of random/structured sensing matrices and dictionaries. It
-also provides some tools for dictionary analysis like:
-coherence, babel function, frame bounds, etc..
-
-The `cr.sparse.cluster` package includes JAX versions of 
-K-means clustering, spectral clustering. It further
-includes support for Sparse Subspace Clustering (SSC) where 
-data points are assumed to lie in a small set of disjoint low dimensional
-subspaces of an ambient space. Currently SSC is implemented using
-OMP. 
-
+systems.
+These algorithms can work with unstructured random and dense sensing matrices
+as well as structured sensing matrices represented as linear operators
 The `cr.sparse.lop` package includes a collection of linear operators
-influenced by `PyLops` [@ravasi2019pylops]. The design is different
-following functional programming principles.
-It is possible to combine different operators  
-to form new operators. 
-
+influenced by `PyLops` [@ravasi2019pylops]. 
 `cr.sparse.wt` package includes a JAX version of major functionality
 from `PyWavelets` [@lee2019pywavelets] making it a first major pure 
 Python wavelets implementation which can work across CPUs, GPUs and TPUs.
 
-The `cr.sparse.dsp` package includes JAX versions of 
-`DCT`, `IDCT`, `WHT` transforms. It also provides a number of utilities
-to construct synthetic signals like chirps, pulses, Gaussian pulses,
-etc..
-
-The `cr.sparse.la` package provides JAX versions of a set of linear algebra
-subroutines as required by other higher level modules. 
-It also includes special routines for solving truncated SVD problems.
 
 # Statement of need
 
@@ -185,77 +150,44 @@ speech recognition, seismology, direction of arrival.
 
 # Sparse signal processing problems and available solvers
 
-We provide JAX based implementations for the following greedy pursuit algorithms:
+We provide JAX based implementations for the following algorithms:
 
[…]
