### 2021-12-02 #3917 @pdebuyl
Hi @shailesh1729

I went through the papers and have minor comments:
- Can you archive the companion https://carnotresearch.github.io/cr-sparse-companion/intro.html on zenodo (the rendered version or rendered+source) and cite it in your article? It is supplemental data to the code so it is better if it is archived.
- For "Yang, J., & Zhang, Y. (2011)", "Kim, S.-J., Koh, K., Lustig, M., Boyd, S., & Gorinevsky, D. (2007)", the title of the article is missing.
- Please add the link to "yall user guide" (Zhang, Yang & Yin 2010) https://www.caam.rice.edu/~optimization/L1/YALL1/User_Guide/YALL1v1.0_User_Guide.pdf (or the 2009 version if it is what you used, I picked the latest from the code's webpage).
- Please add the preprint URL for Ravasi, M., & Vasconcelos, I. (2019).
- First sentence, recommended wording: "We introduceCR-Sparse, a Python library that enables efficiently solving a wide variety of6sparse representation based signal processing problems" -> "We introduce CR-Sparse, a Python library that enables to efficiently solve a wide variety of sparse representation based signal processing problems".
- Line 50: add a space after link to YALL1

If the code has not changed since release v0.2.1 https://zenodo.org/record/5635717 I'll use that as the archive, can you confirm?

The new version of the article is much better in terms of length, thanks!
