### 2021-12-02 1999c29
paper updates on feedback by pdebuyl
