+++ paper/paper.bib
@@ -286,7 +286,7 @@
 }
 
 @article{kim2007interior,
-  title={An interior-point method for large-scale $$\backslash$ell\_1 $-regularized least squares},
+  title={An interior-point method for large-scale l1-regularized least squares},
   author={Kim, Seung-Jean and Koh, Kwangmoo and Lustig, Michael and Boyd, Stephen and Gorinevsky, Dimitry},
   journal={IEEE journal of selected topics in signal processing},
   volume={1},
@@ -453,8 +453,19 @@ address = {Natick, Massachusetts}
   title={PyLops--A Linear-Operator Python Library for large scale optimization},
   author={Ravasi, Matteo and Vasconcelos, Ivan},
   journal={arXiv preprint arXiv:1907.12349},
-  year={2019}
+  year={2019},
+  url={https://arxiv.org/abs/1907.12349}
 }
+
+@software{shailesh2021companion,
+  author = {Kumar, Shailesh},
+  title = {CR-Sparse Companion},
+  doi={10.5281/zenodo.5749656},
+  url = {https://doi.org/10.5281/zenodo.5749656},
+  version = {0.2.1-docs},
+  year = {2021},
+}
+
 
 @article{simpson2021classo,
   doi = {10.21105/joss.02844},
@@ -530,7 +541,7 @@ address = {Natick, Massachusetts}
 }
 
 @article{yang2011alternating,
-  title={Alternating direction algorithms for l_1-problems in compressive sensing},
+  title={Alternating direction algorithms for l1-problems in compressive sensing},
   author={Yang, Junfeng and Zhang, Yin},
   journal={SIAM journal on scientific computing},
   volume={33},
@@ -561,9 +572,10 @@ address = {Natick, Massachusetts}
 
 
 @techreport{zhang2009user,
-  title={User's guide for yall1: Your algorithms for l1 optimization},
-  author={Zhang, Yin},
-  year={2009}
+  title={User's guide for YALL1: Your algorithms for L1 optimization},
+  author={Zhang, Yin and Yang, Junfeng and Yin, Wotao},
+  year={2010},
+  url={https://www.caam.rice.edu/~optimization/L1/YALL1/User_Guide/YALL1v1.0_User_Guide.pdf}
 }
 
 @article{zhang2012hybrid,
