+++ _sources/gallery/lop/wt_op.rst.txt
@@ -113,7 +113,7 @@ Sampling frequency
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf197bdc0>]
+    [<matplotlib.lines.Line2D object at 0x7fd19b225a00>]
 
 
 
@@ -164,7 +164,7 @@ Wavelet coefficients
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf12104c0>]
+    [<matplotlib.lines.Line2D object at 0x7fd19a94b8b0>]
 
 
 
@@ -199,7 +199,7 @@ Let's keep only 10 percent of the coefficients
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf11e1a60>]
+    [<matplotlib.lines.Line2D object at 0x7fd19a920e50>]
 
 
 
@@ -244,7 +244,7 @@ Reconstruction
     0.034259365997888334
     36.56352988951204
 
-    <matplotlib.legend.Legend object at 0x7faaf10c67c0>
+    <matplotlib.legend.Legend object at 0x7fd19a805bb0>
 
 
 
@@ -369,7 +369,7 @@ Reconstruction
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  9.718 seconds)
+   **Total running time of the script:** ( 0 minutes  10.093 seconds)
 
 
 .. _sphx_glr_download_gallery_lop_wt_op.py:
