+++ _sources/gallery/rec_l1/partial_wh_sensor_cosine_basis.rst.txt
@@ -113,7 +113,7 @@ Non-sparse signal
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf0c7bfa0>]
+    [<matplotlib.lines.Line2D object at 0x7fd19a291820>]
 
 
 
@@ -149,7 +149,7 @@ The Sparsifying Basis
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf08c0c40>]
+    [<matplotlib.lines.Line2D object at 0x7fd19a94b490>]
 
 
 
@@ -234,7 +234,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf0b24370>]
+    [<matplotlib.lines.Line2D object at 0x7fd199fc1f70>]
 
 
 
@@ -307,14 +307,14 @@ Solution
  .. code-block:: none
 
 
-    <matplotlib.legend.Legend object at 0x7faaf058f5e0>
+    <matplotlib.legend.Legend object at 0x7fd199cc7fd0>
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  7.146 seconds)
+   **Total running time of the script:** ( 0 minutes  7.417 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_partial_wh_sensor_cosine_basis.py:
