+++ _sources/gallery/rec_l1/spikes_l1ls.rst.txt
@@ -123,7 +123,7 @@ The Spikes
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf1796460>]
+    [<matplotlib.lines.Line2D object at 0x7fd199f89550>]
 
 
 
@@ -181,7 +181,7 @@ Measurement process
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf084ebb0>]
+    [<matplotlib.lines.Line2D object at 0x7fd19a0c71c0>]
 
 
 
@@ -267,7 +267,7 @@ Solution
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf0483ca0>]
+    [<matplotlib.lines.Line2D object at 0x7fd199bc36d0>]
 
 
 
@@ -299,7 +299,7 @@ The magnitudes of non-zero values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf0464100>]
+    [<matplotlib.lines.Line2D object at 0x7fd199b95af0>]
 
 
 
@@ -335,7 +335,7 @@ Thresholding for large values
  .. code-block:: none
 
 
-    [<matplotlib.lines.Line2D object at 0x7faaf037e790>]
+    [<matplotlib.lines.Line2D object at 0x7fd199abe1c0>]
 
 
 
@@ -416,14 +416,14 @@ Improvement using least squares over support
     (1024, 160)
     ls_rel_error=2.057e-02
 
-    [<matplotlib.lines.Line2D object at 0x7faaf02b08e0>]
+    [<matplotlib.lines.Line2D object at 0x7fd199973310>]
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  8.132 seconds)
+   **Total running time of the script:** ( 0 minutes  8.222 seconds)
 
 
 .. _sphx_glr_download_gallery_rec_l1_spikes_l1ls.py:
