### 2021-12-02 #3917 @shailesh1729
HI @pdebuyl,

Thank you for the detailed feedback. I have corrected the bibliographic and other editing issues in the new draft.

I have created the zenodo archive for the companion site:
- code (notebooks) from main branch https://doi.org/10.5281/zenodo.5749633
- rendered Jupyter book version from the gh-pages branch https://doi.org/10.5281/zenodo.5749656

I have cited the second one in the paper.

During the last month of review, while the code didn't change, the documentation and README were revised significantly based on the feedback received here.  Hence, I have created a new release (v0.2.2) just now capturing all the doc updates.

The Zenodo archive link for v0.2.2 is:  https://doi.org/10.5281/zenodo.5749792
This should be used for archival purposes.

Hope this addresses the concerns listed by you.

Thank you again for all your help in improving the quality of the software and the paper.
