### 2021-12-02 #3917 @kthyng
Hi @shailesh1729! I'll be helping with final publishing.

Can you update the metadata on your Zenodo archive so the title and author list exactly match your JOSS paper?
