### 2021-12-02 #3917 @pdebuyl
@kthyng "metadata on your Zenodo archive so the title and author list exactly match your JOSS paper" -> sorry, I forgot to check it
