### 2021-12-02 #3917 @pdebuyl
and the LSQR entry from the TOMS does not have a DOI to avoid you looking for it :-)
