### 2021-12-02 9546bb7
missing doi for LSQR
