+++ paper/paper.bib
@@ -383,6 +383,8 @@ address = {Natick, Massachusetts}
   number={1},
   pages={43--71},
   year={1982},
+  doi={10.1145/355984.355989},
+  url={https://doi.org/10.1145/355984.355989},
   publisher={ACM New York, NY, USA}
 }
 
