### 2021-12-02 #3917 @shailesh1729
It was my mistake. I added the LSQR reference later and forgot to add the DOI.
