### 2021-12-02 #3917 @shailesh1729
All DOIs are good now in the updated draft.
